vlc_project/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
reedsolo>=1.7.0
Pillow>=10.0.0
